+++ src/utils/index.js
@@ -1,7 +1,11 @@
 // @flow
 
-export default function kebabify(prop: string): string {
-  const upperToHyphen = (match: string, offset: number, string: string): string => {
+export function kebabify(prop: string): string {
+  const upperToHyphen = (
+    match: string,
+    offset: number,
+    string: string
+  ): string => {
     const addDash = offset && string.charAt(offset - 1) !== '-';
 
     return (addDash ? '-' : '') + match.toLowerCase();
@@ -9,3 +13,6 @@ export default function kebabify(prop: string): string {
 
   return prop.replace(/[A-Z]/g, upperToHyphen);
 }
+
+export const isPlainObject = (arg: any) =>
+  Object.prototype.toString.call(arg) === '[object Object]';