+++ test/utils.spec.js
@@ -1,5 +1,4 @@
-import kebabify from '../src/utils';
-
+import { kebabify, isPlainObject } from '../src/utils';
 
 describe('kekabify', () => {
   it('should convert camelCase properties to kebab-case', () => {
@@ -20,5 +19,14 @@ describe('kekabify', () => {
     const result = input.map(kebabify);
 
     expect(result).toMatchSnapshot();
+  });
+});
+
+describe('isPlainObject', () => {
+  it('should assert for plain object types', () => {
+    expect(isPlainObject({})).toBe(true);
+    expect(isPlainObject([])).toBe(false);
+    expect(isPlainObject(null)).toBe(false);
+    expect(isPlainObject('')).toBe(false);
   });
 });